Elifdev2004

+++ src/app/modules/admin/admin-user-management/admin-user-management.component.ts
@@ -1,10 +1,8 @@
 import { Component } from '@angular/core';
 import { User } from '../../../shared/dto/user';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ProductService } from '../../../shared/service/product.service';
 import { ToastrService } from 'ngx-toastr';
 import { UserService } from '../../../shared/service/user.service';
-import { FormControl, FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-admin-user-management',
@@ -15,7 +13,6 @@ export class AdminUserManagementComponent {
   users: User[] = [];
   searchText: string = ''; // Arama metni için değişken eklendi
   selectedUser: User | null = null;
-  // myGroup: FormGroup;
 
   constructor(
     private router: Router,
@@ -23,30 +20,13 @@ export class AdminUserManagementComponent {
     private userService: UserService,
     private toastr: ToastrService,
   ) { 
-    // this.myGroup = new FormGroup({
-    //   searchText: new FormControl()
-    // });
   }
   //Component çağrıldığında çalışan method.
   ngOnInit(): void {
-    //ProductService'den getAllProducts() methodunu çağrıyor. Tüm productları döndürüyor.
-    // this.userService.getAllUsers().subscribe({
-    //   next: (users => {
-    //     console.log(users);
-    //     this.users = users;
-    //   })
-    // });
     this.loadUsers();
   }
 
   loadUsers(): void {
-    // ProductService'den getAllProducts() methodunu çağrıyor. Tüm productları döndürüyor.
-    // this.userService.getAllUsers().subscribe({
-    //   next: (users => {
-    //     console.log(users);
-    //     this.users = users;
-    //   })
-    // });
     this.userService.getAllUsers().subscribe(
       (data: User[]) => {
         this.users = data;
@@ -55,6 +35,7 @@ export class AdminUserManagementComponent {
   }
 
   addUser(){
+    this.userService.editingUser = null;
     this.router.navigate(['addUser'], { relativeTo: this.route });
   }
   editUser(user: User) {
@@ -69,6 +50,12 @@ export class AdminUserManagementComponent {
       console.error("No user selected for deletion !");
       return;
     }
+    const loggedInUserEmail = this.userService.getloggedInUserEmail();
+    if (loggedInUserEmail === user.email) {
+      this.toastr.error("Can Not Delete !")
+      return
+    }
+
     this.userService.deleteUser(user).subscribe({
       next: () => {
         this.users = this.users.filter(u => u !== user);
@@ -88,8 +75,5 @@ export class AdminUserManagementComponent {
       ); 
     }
     return filteredUsers;
-    // return this.users.filter(user => {
-    //   return user.email.toLowerCase().includes(this.searchText.toLowerCase());
-    // });
   }
 }
